fix: split cloudflare api implementation into a service
